extend toolbar to full width of screen, to prepare for alternate editor layouts
